+++ src/mrt.c
@@ -765,6 +765,7 @@ static mrtentry_t *alloc_mrtentry(srcentry_t *src, grpentry_t *grp)
     PIMD_VIFM_CLRALL(mrt->sg_joined_oifs);
     PIMD_VIFM_CLRALL(mrt->leaves);
     PIMD_VIFM_CLRALL(mrt->pruned_oifs);
+    PIMD_VIFM_CLRALL(mrt->prune_pending_oifs);
     PIMD_VIFM_CLRALL(mrt->asserted_oifs);
     PIMD_VIFM_CLRALL(mrt->oifs);
     mrt->upstream = NULL;
@@ -780,16 +781,14 @@ static mrtentry_t *alloc_mrtentry(srcentry_t *src, grpentry_t *grp)
      */
 #ifdef SAVE_MEMORY
     mrt->vif_timers	    = calloc(1, sizeof(uint16_t) * numvifs);
-    mrt->vif_deletion_delay = calloc(1, sizeof(uint16_t) * numvifs);
     mrt->asserts	    = calloc(numvifs, sizeof(mrt->asserts[0]));
     vif_numbers = numvifs;
 #else
     mrt->vif_timers	    = calloc(1, sizeof(uint16_t) * total_interfaces);
-    mrt->vif_deletion_delay = calloc(1, sizeof(uint16_t) * total_interfaces);
     mrt->asserts	    = calloc(total_interfaces, sizeof(mrt->asserts[0]));
     vif_numbers = total_interfaces;
 #endif /* SAVE_MEMORY */
-    if (!mrt->vif_timers || !mrt->vif_deletion_delay || !mrt->asserts) {
+    if (!mrt->vif_timers || !mrt->asserts) {
 	logit(LOG_WARNING, 0, "alloc_mrtentry(): out of memory");
 	FREE_MRTENTRY(mrt);
 	return NULL;
@@ -797,9 +796,6 @@ static mrtentry_t *alloc_mrtentry(srcentry_t *src, grpentry_t *grp)
 
     /* Reset the timers */
     for (i = 0, timer = mrt->vif_timers; i < vif_numbers; i++, timer++)
-	RESET_TIMER(*timer);
-
-    for (i = 0, timer = mrt->vif_deletion_delay; i < vif_numbers; i++, timer++)
 	RESET_TIMER(*timer);
 
     mrt->flags = MRTF_NEW;